requirements.txt
flask==1.1.2
flask-sqlalchemy==2.4.4
#flask-marshmallow==0.14.0
python-dotenv==0.15.0
Flask-Migrate==2.6.0
psycopg2==2.8.6
hashlib